I slightly changed the function names in the timing library so that they are a little more descriptive. Didn't change any of the functionality. I moved the planner include files into planner.h. I changed visiond to use the new timing library. I modified the visiond helper functions to have visiond_ prepended to the function names. I also continued renaming makefile to Makefile when I see it just because it seems to be a more standard way of naming. If you do want to rename them use 'svn mv makefile Makefile' or else it might cause problems.
!!! WARNING !!!
I don't know if visiond was compiling before but it's not now. Sorry! It is
complaining that it can't find functions in the vision library. I've tried
changing the order that include files are listed in visiond.c but can't get it
right. Sometimes it will not complain about the vision functions but instead
will say that MSG_DATA is not defined.


git-svn-id: file:///home/thomas/tmp/stingray/libs@269 ae4af862-681a-43e1-b28c-8c040b23e139

+++ parser/include/parser.h
@@ -177,9 +177,9 @@ typedef struct _CONF_VARS {
 	int			save_image_bottom;
 	int			save_image_color;
 	int			save_image_binary;
-	int			save_image_rate;
+	float		save_image_rate;
 	char		save_image_dir[STRING_SIZE];
-	int			open_image_rate;
+	float		open_image_rate;
 	char		open_image_dir[STRING_SIZE];
 	float		task_init_yaw;
 } CONF_VARS;
